# api/openapi/business-api.yaml
openapi: 3.0.3
info:
  title: Dora Business API
  version: '2026-06-27'
  description: Production M0 HTTP contract for user, public and admin business service APIs. Every operation uses named request
    and response DTOs; no placeholder body or generic response components are allowed.
servers:
- url: http://localhost:19080
...
paths:
  /api/auth/register:
    post:
      tags:
      - Auth
      operationId: registerPersonalAccount
      responses:
        '200':
          $ref: '#/components/responses/RegisterPersonalAccountOk'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/AuthProblem'
        '403':
          $ref: '#/components/responses/PermissionProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '500':
          $ref: '#/components/responses/SystemProblem'
      requestBody:
        $ref: '#/components/requestBodies/RegisterPersonalAccountBody'
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
  /api/auth/login:
    post:
      tags:
      - Auth
      operationId: login
      responses:
        '200':
          $ref: '#/components/responses/LoginOk'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/AuthProblem'
        '403':
          $ref: '#/components/responses/PermissionProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '500':
          $ref: '#/components/responses/SystemProblem'
      requestBody:
        $ref: '#/components/requestBodies/LoginBody'
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
  /api/auth/logout:
    post:
      tags:
      - Auth
      operationId: logout
      responses:
        '200':
          $ref: '#/components/responses/LogoutOk'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/AuthProblem'
        '403':
          $ref: '#/components/responses/PermissionProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '500':
          $ref: '#/components/responses/SystemProblem'
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        $ref: '#/components/requestBodies/LogoutBody'
  /api/account/current-space:
    get:
      tags:
      - Account
      operationId: getCurrentSpace
      responses:
        '200':
          $ref: '#/components/responses/GetCurrentSpaceOk'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/AuthProblem'
        '403':
          $ref: '#/components/responses/PermissionProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '500':
          $ref: '#/components/responses/SystemProblem'
  /api/account/switch-identity:
    post:
      tags:
      - Account
      operationId: switchIdentity
      responses:
        '200':
          $ref: '#/components/responses/SwitchIdentityOk'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/AuthProblem'
        '403':
          $ref: '#/components/responses/PermissionProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '500':
          $ref: '#/components/responses/SystemProblem'
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        $ref: '#/components/requestBodies/SwitchIdentityBody'
  /api/enterprise/register:
    post:
      tags:
      - Enterprise
      operationId: createEnterprise
      responses:
        '200':
          $ref: '#/components/responses/CreateEnterpriseOk'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/AuthProblem'
        '403':
          $ref: '#/components/responses/PermissionProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '500':
          $ref: '#/components/responses/SystemProblem'
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        $ref: '#/components/requestBodies/CreateEnterpriseBody'
  /api/enterprise/summary:
    get:
      tags:
      - Enterprise
      operationId: getEnterpriseSummary
      responses:
        '200':
          $ref: '#/components/responses/GetEnterpriseSummaryOk'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/AuthProblem'
        '403':
          $ref: '#/components/responses/PermissionProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '500':
          $ref: '#/components/responses/SystemProblem'
  /api/enterprise/members:
    get:
      tags:
      - Enterprise
      operationId: listEnterpriseMembers
      responses:
        '200':
          $ref: '#/components/responses/ListEnterpriseMembersOk'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/AuthProblem'
        '403':
          $ref: '#/components/responses/PermissionProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '500':
          $ref: '#/components/responses/SystemProblem'
      parameters:
      - $ref: '#/components/parameters/PageSize'
      - $ref: '#/components/parameters/PageToken'
  /api/enterprise/invites:
    post:
      tags:
      - Enterprise
      operationId: createEnterpriseInvite
      responses:
        '200':
          $ref: '#/components/responses/CreateEnterpriseInviteOk'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/AuthProblem'
        '403':
          $ref: '#/components/responses/PermissionProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '500':
          $ref: '#/components/responses/SystemProblem'
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        $ref: '#/components/requestBodies/CreateEnterpriseInviteBody'
  /api/enterprise/members/{member_id}/remove/preview:
    post:
      tags:
      - Enterprise
      operationId: previewRemoveEnterpriseMember
      responses:
        '200':
          $ref: '#/components/responses/PreviewRemoveEnterpriseMemberOk'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/AuthProblem'
        '403':
          $ref: '#/components/responses/PermissionProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '500':
          $ref: '#/components/responses/SystemProblem'
      parameters:
      - $ref: '#/components/parameters/MemberId'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        $ref: '#/components/requestBodies/PreviewRemoveEnterpriseMemberBody'
  /api/enterprise/members/{member_id}/remove/confirm:
    post:
      tags:
      - Enterprise
      operationId: confirmRemoveEnterpriseMember
      responses:
        '200':
          $ref: '#/components/responses/ConfirmRemoveEnterpriseMemberOk'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/AuthProblem'
        '403':
          $ref: '#/components/responses/PermissionProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '500':
          $ref: '#/components/responses/SystemProblem'
      parameters:
      - $ref: '#/components/parameters/MemberId'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        $ref: '#/components/requestBodies/ConfirmRemoveEnterpriseMemberBody'
  /api/enterprise/owner-transfer/preview:
    post:
      tags:
      - Enterprise
      operationId: previewTransferOwner
      responses:
        '200':
          $ref: '#/components/responses/PreviewTransferOwnerOk'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/AuthProblem'
        '403':
          $ref: '#/components/responses/PermissionProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '500':
          $ref: '#/components/responses/SystemProblem'
      requestBody:
        $ref: '#/components/requestBodies/PreviewTransferOwnerBody'
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
  /api/enterprise/owner-transfer/confirm:
    post:
      tags:
      - Enterprise
      operationId: confirmTransferOwner
      responses:
        '200':
          $ref: '#/components/responses/ConfirmTransferOwnerOk'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/AuthProblem'
        '403':
          $ref: '#/components/responses/PermissionProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '500':
          $ref: '#/components/responses/SystemProblem'
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        $ref: '#/components/requestBodies/ConfirmTransferOwnerBody'
  /api/projects:
    get:
      tags:
      - Account
      operationId: listProjects
      responses:
        '200':
          $ref: '#/components/responses/ListProjectsOk'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/AuthProblem'
        '403':
          $ref: '#/components/responses/PermissionProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '500':
          $ref: '#/components/responses/SystemProblem'
      parameters:
      - $ref: '#/components/parameters/PageSize'
      - $ref: '#/components/parameters/PageToken'
    post:
      tags:
      - Account
      operationId: createProject
      responses:
        '200':
          $ref: '#/components/responses/CreateProjectOk'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/AuthProblem'
        '403':
          $ref: '#/components/responses/PermissionProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '500':
          $ref: '#/components/responses/SystemProblem'
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        $ref: '#/components/requestBodies/CreateProjectBody'
  /api/projects/{project_id}:
    get:
      tags:
      - Account
      operationId: getProject
      responses:
        '200':
          $ref: '#/components/responses/GetProjectOk'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/AuthProblem'
        '403':
          $ref: '#/components/responses/PermissionProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '500':
          $ref: '#/components/responses/SystemProblem'
      parameters:
      - $ref: '#/components/parameters/ProjectId'
    patch:
      tags:
      - Account
      operationId: updateProject
      responses:
        '200':
          $ref: '#/components/responses/UpdateProjectOk'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/AuthProblem'
        '403':
          $ref: '#/components/responses/PermissionProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '500':
          $ref: '#/components/responses/SystemProblem'
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        $ref: '#/components/requestBodies/UpdateProjectBody'
  /api/projects/{project_id}/archive:
    post:
      tags:
      - Account
      operationId: archiveProject
      responses:
        '200':
          $ref: '#/components/responses/ArchiveProjectOk'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/AuthProblem'
        '403':
          $ref: '#/components/responses/PermissionProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '500':
          $ref: '#/components/responses/SystemProblem'
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        $ref: '#/components/requestBodies/ArchiveProjectBody'
  /api/projects/{project_id}/restore:
    post:
      tags:
      - Account
      operationId: restoreProject
      responses:
        '200':
          $ref: '#/components/responses/RestoreProjectOk'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/AuthProblem'
        '403':
          $ref: '#/components/responses/PermissionProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '500':
          $ref: '#/components/responses/SystemProblem'
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        $ref: '#/components/requestBodies/RestoreProjectBody'
  /api/projects/{project_id}/assets:
    get:
      tags:
      - Asset
      operationId: listProjectAssets
      responses:
        '200':
          $ref: '#/components/responses/ListProjectAssetsOk'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/AuthProblem'
        '403':
          $ref: '#/components/responses/PermissionProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '500':
          $ref: '#/components/responses/SystemProblem'
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/PageSize'
  /api/projects/{project_id}/works:
    get:
      tags:
      - Work
      operationId: listProjectWorks
      responses:
        '200':
          $ref: '#/components/responses/ListProjectWorksOk'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/AuthProblem'
        '403':
          $ref: '#/components/responses/PermissionProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '500':
          $ref: '#/components/responses/SystemProblem'
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/PageSize'
...
  /api/admin/auth/login:
    post:
      tags:
      - Admin
      operationId: adminLogin
      responses:
        '200':
          $ref: '#/components/responses/AdminLoginOk'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/AuthProblem'
        '403':
          $ref: '#/components/responses/PermissionProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '500':
          $ref: '#/components/responses/SystemProblem'
      requestBody:
        $ref: '#/components/requestBodies/AdminLoginBody'
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
...
  /api/admin/dashboard:
    get:
      tags:
      - Admin
      operationId: getAdminDashboard
      responses:
        '200':
          $ref: '#/components/responses/GetAdminDashboardOk'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '401':
          $ref: '#/components/responses/AuthProblem'
        '403':
          $ref: '#/components/responses/PermissionProblem'
        '409':
          $ref: '#/components/responses/ConflictProblem'
        '500':
          $ref: '#/components/responses/SystemProblem'
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        maxLength: 128
...
    PageToken:
      name: page_token
      in: query
      schema:
        type: string
...
  requestBodies:
...
    RegisterPersonalAccountBody:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RegisterPersonalAccountRequest'
...
  responses:
...
    ValidationProblem:
      description: Invalid request fields.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    AuthProblem:
      description: Unauthenticated request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    PermissionProblem:
      description: Permission denied.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ConflictProblem:
      description: Business state or idempotency conflict.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    SystemProblem:
      description: System error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    RegisterPersonalAccountOk:
      description: registerPersonalAccount successful response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RegisterPersonalAccountOkEnvelope'
...